% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow.R
\docType{data}
\name{get_metric}
\alias{get_metric}
\title{Get common metrics' formulas}
\format{
An object of class \code{list} of length 2.
}
\usage{
get_metric
}
\description{
Get common metrics' formulas
}
\keyword{datasets}
