% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyopt-package.R
\docType{package}
\name{phyopt-package}
\alias{phyopt}
\alias{phyopt-package}
\title{phyopt: General-Use Population-Based Hybrid Optimization}
\description{
todo.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://https://github.com/ricardo-semiao/phyopt}
  \item \url{https://ricardo-semiao.github.io/phyopt/}
  \item Report bugs at \url{https://github.com/ricardo-semiao/phyopt/issues}
}

}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
