% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyopt-package.R
\docType{package}
\name{phyopt-package}
\alias{phyopt}
\alias{phyopt-package}
\title{phyopt: General-Use Population-Based Hybrid Optimization}
\description{
todo.
}
\author{
\strong{Maintainer}: Ricardo Semião \email{ricardo.semiao@outlook.com}

}
\keyword{internal}
